backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
firebase-admin>=6.2.0
pydantic>=2.0.0
python-dotenv>=0.19.0,<0.20.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
facebook-sdk>=3.1.0
google-api-python-client>=2.97.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
openai>=0.27.8
scikit-learn>=1.3.0
pandas>=2.0.3
numpy>=1.24.3
requests>=2.31.0
aiohttp>=3.8.5
python-firebase>=1.2.1
httpx>=0.18.1,<0.19.0
python-facebook-api>=0.16.0
instagram-private-api>=1.6.0
alembic>=1.6.5,<1.7.0
sqlalchemy>=1.4.0,<1.5.0
psycopg2-binary>=2.9.1,<3.0.0
tenacity>=8.0.1,<8.1.0
emails>=0.6,<0.7
raven>=6.10.0
gunicorn>=21.2.0
google-ads>=22.1.0
python-facebook-graph-api>=0.10.1
instabot>=0.117.0
google-cloud-storage>=2.10.0
stripe>=2.60.0,<2.61.0
websockets>=10.0,<11.0
redis>=4.0.0,<4.1.0
aioredis>=2.0.0,<2.1.0
prometheus-client>=0.11.0,<0.12.0
sentry-sdk>=1.3.1,<1.4.0
jinja2>=3.0.1,<3.1.0
